Updated Variable Calls and repository items
This push *should* for the most part clear up some issues. Currently in the process of fixing up main.lua as it was having trouble reading the table of a file that doesn't yet exist (nor does variables in said table get set yet). This should be rectified in the next push.

I will handle the main.lua so for now don't modify or push it to master.

-WildRyce