V0.1.1 Push
Base Variables updated push. This fixes some of the major issues tthat stopped the game from running. If you encounter a crash, please take note of the line and script it was from (the crash shows). If you can fix it (say it was missing a "Variables." in front of a variable) feel free to do so and push it to the master. Else just take note of it and pass it along.

+ Fixed button size in inventory for Inspire potions (was previously able to clip off the right side of the screen).
+ Fixed variable names and updated all scripts to account for the new save/load data handler, and new dictionary (table) handler.
- Removed a few unused variables that would no longer serve any purpose.
- Removed old save/load data functions.
- Removed GlobalData.lua. The script was no longer needed (all base variables were shifted to main.lua to actually handle the new save/load data, and just shifted over monsters data as it had no problems swapping over).