Handlers Push
This push fixes some issues regarding the level-up system. It was quite glitchy beforehand, but this push looks to fix said issues. Minor issue to be fixed where player level and health do not update on dungeon screen when staying on the dungeon screen 2+ levelups in a row. Swapping scenes updates it when going back. Reason is that the Update() function stop firing after the overlay shows up.

For more info check the readme.txt for more detailed.